refactor: ContributionGraph 컴포넌트 구조 개선
- 목표가 없을 때 표시되지 않도록 개선
